+++ contracts/tokenwrapper/src/utils.rs
@@ -30,12 +30,12 @@ pub fn is_valid_unwrap_amount(deps: DepsMut, sender: &str, amount: Uint128) -> b
 }
 
 // Calculates the "fee" from "wrap_amount".
-pub fn get_fee_from_amount(amount_to_wrap: Uint128, fee_perc: u8) -> Uint128 {
+pub fn get_fee_from_amount(amount_to_wrap: Uint128, fee_perc: u16) -> Uint128 {
     amount_to_wrap.multiply_ratio(fee_perc, WRAP_FEE_CALC_DENOMINATOR)
 }
 
 // Calculate the "amount_to_send" from "deposit_target" amount.
-pub fn get_amount_to_wrap(target_amount: Uint128, fee_perc: u8) -> Uint128 {
+pub fn get_amount_to_wrap(target_amount: Uint128, fee_perc: u16) -> Uint128 {
     target_amount.multiply_ratio(
         WRAP_FEE_CALC_DENOMINATOR,
         WRAP_FEE_CALC_DENOMINATOR - fee_perc,
